use WKT instead of SVG + fix orcid bug
In this commit we transition from SVG to WKT (well known text) to encode
polygons and points. This commit also fixes a bug in which adding an
orcid id would raise an error (an item was missing in the rdf triple).
We use schema.org/identifier as the type for the orcid id.